decoupled the prisma client from server.js, as it was stopping me from updating the staging database. now started in a separate client.js file
