+++ backend/server.js
@@ -4,7 +4,8 @@ import express from 'express';
 import cors from 'cors';
 import session from 'express-session';
 import userRoutes from './routes/userRoutes.js';
-import { PrismaClient } from "@prisma/client";
+// import { PrismaClient } from "@prisma/client";
+import { prisma } from "./prisma/client.js"
 import adminRoutes from "./routes/adminRoutes.js";
 import authorRoutes from "./routes/authorRoutes.js";
 import superAdminRoutes from "./routes/superAdminRoutes.js";
@@ -13,9 +14,9 @@ import { fileURLToPath } from 'url';
 
 dotenv.config();
 const app = express();
-export const prisma = new PrismaClient({
-  log: ['error'],
-});
+// export const prisma = new PrismaClient({
+//   log: ['error'],
+// });
 // export const redisClient = redis.createClient();
 // redisClient.on("error", (err) => console.error("Redis Error:", err));
 // redisClient.connect().then(() => console.log("Connected to Redis"));
